requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic
pillow
opencv-python
numpy
